+++ src/boxTemplate.h
@@ -11,7 +11,7 @@ namespace engix
         BoxTemplate(const assets::Path& path) : BoxTemplate(load(path)) {}
         
         static BoxTemplate load(const assets::Path& path);
-        Texture genTexture(size_t width, size_t height) const;
+        Texture genTexture(int width, int height) const;
 
         PixelImage left, up, right, down;
         PixelImage leftUp, rightUp, leftDown, rightDown;
